% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeStamp.R
\name{fixTimeStamp}
\alias{fixTimeStamp}
\alias{selezionaIntervalloTimeStamp}
\title{Funzioni di gestione delle date}
\usage{
fixTimeStamp(db, campoData = "created", timeRange = NULL)
}
\arguments{
\item{db}{data.frame contenente i tweets.}

\item{campoData}{"created" e "ts" sono due campi data del db estratto da
dump_twitter.R}

\item{timeRange}{due valori di tipo data indicanti inizio e fine.}
}
\value{
l'output e' db "aggiustato"
}
\description{
Funzioni di gestione delle date
}
\details{
aggiungere dettagli qui
}
\note{
%% ~~further notes~~
}
\examples{

 \dontrun{TW=fixTimeStamp(TW)}
 \dontrun{TW=selezionaIntervallo(TW,as.POSIXct(c("2013-12-27 17:54:42 CET", "2013-12-27 22:33:38 CET")))}
 \dontrun{TW$created.round <- as.POSIXct(round(t$created,"hour"))}



}
\author{
Dario Solari, Livio Finos
}
\keyword{~kwd1}
\keyword{~kwd2}
