% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizzaemote.R
\name{normalizzaemote}
\alias{normalizzaemote}
\title{normalizzaemote}
\usage{
normalizzaemote(testo, perl = TRUE)
}
\arguments{
\item{testo}{a set of texts to be preprocessed}

\item{perl}{logical. If TRUE Perl-compatible regexps are used.}
}
\value{
a set of text where emoticons has been replaced by the respective keyword
}
\description{
\code{normalizzaemote} replaces emoticons in \code{testo} with a keyword corresponding the feeling described
}
\author{
Mattia Da Pont, Livio Finos, Dario Solari
}
