% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizzaEmoticons.R
\name{normalizzaEmoticons}
\alias{normalizzaEmoticons}
\title{normalizzaemote}
\usage{
normalizzaEmoticons(testo)
}
\arguments{
\item{testo}{a set of texts to be preprocessed}
}
\value{
a set of text where emoticons has been replaced by the respective keyword
}
\description{
Identifies the UTF-8 codes of emoticons in the given text and replaces them with a corresponding keyword
}
\details{
\code{normalizzaEmoticons} replaces emoticons in \code{testo} with a keyword corresponding to what is represented
}
\author{
Livio Finos, Antonio Calcagnì, Mattia Da Pont
}
