% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizzahtml.R
\name{normalizzahtml}
\alias{normalizzahtml}
\title{normalizzahtml}
\usage{
normalizzahtml(testo, perl = TRUE, fixed = TRUE)
}
\arguments{
\item{testo}{a set of texts to be preprocessed}

\item{perl}{logical. If TRUE Perl-compatible regexps are used.}

\item{fixed}{logical. If TRUE, pattern is a string to be matched as is. Overrides all conflicting arguments.}
}
\value{
a set of texts where URLS have been replaced by the keyword WWWURLWWW
}
\description{
\code{normalizzahtml} replaces links and URLs in \code{testo} with the keyword WWWURLWWW
}
\examples{

 testo<-c("http:textwiller.com","www.textwiller.com") 
 normalizzahtml(testo)
 
 
}
\author{
Livio Finos
}
