% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizzapunteggiatura.R
\name{normalizzapunteggiatura}
\alias{normalizzapunteggiatura}
\title{normalizzapunteggiatura}
\usage{
normalizzapunteggiatura(
  testo,
  removeUnderscore = TRUE,
  perl = TRUE,
  fixed = TRUE
)
}
\arguments{
\item{testo}{a set of texts to be stripped of escape and punctation}

\item{removeUnderscore}{logical. If TRUE, underscore characters "_" are wiped from the texts}

\item{perl}{logical. If TRUE Perl-compatible regexps are used.}

\item{fixed}{logical. If TRUE, pattern is a string to be matched as is. Overrides all conflicting arguments.}
}
\value{
a set of processed texts.
}
\description{
\code{normalizzapunteggiatura} strip of escape and punctation from  a set of texts.
}
\examples{

 testo<-c(testo<-c("@ ","@retweet","# ","#ciao")) 
 normalizzapunteggiatura(testo)
 
 
}
\author{
Livio Finos
}
