% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizzaslang.R
\name{normalizzaslang}
\alias{normalizzaslang}
\title{normalizzaslang}
\usage{
normalizzaslang(testo, perl = TRUE)
}
\arguments{
\item{testo}{a set of texts to be normalized}

\item{perl}{logical. If TRUE Perl-compatible regexps are used.}
}
\value{
a set of texts where slang expressions of the italian language have been replaced with a keyword
}
\description{
\code{normalizzaslang} replaces slang expressions of the italian language in \code{testo} with
a keyword linked with the corresponding feeling
}
\examples{

 testo<-c("grandissima","aaa","nun")
 normalizzaslang(testo)
 
 
}
\author{
Livio Finos
}
