% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeStopwords.R
\name{removeStopwords}
\alias{removeStopwords}
\title{removeStopwords}
\usage{
removeStopwords(testo, stopwords = itastopwords)
}
\arguments{
\item{testo}{a vector of characters to be normalized}
}
\value{
a vector of normalized characters
}
\description{
normalize the given text by finding and removing stopwords of the italian language 
(listed in itastopwords)
}
\details{
\code{removeStopwords} removes stopwords of the italian language in \code{testo}
}
\examples{
testo<-("mangio una mela e una banana che ho appena comprato")
removeStopwords(testo)
 
}
\author{
Livio Finos
}
