% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{sentiment}
\alias{sentiment}
\alias{sentimentVocabularies}
\alias{vocabolariMadda}
\title{Performs sentiment analysis}
\usage{
sentiment(
  text,
  algorithm = "Maddalena",
  vocabularies = NULL,
  normalizzaTesti = TRUE,
  ...
)
}
\arguments{
\item{text}{descr .}

\item{algorithm}{descr.}

\item{vocabularies}{\code{vocabolariMadda} by default.}

\item{normalizzaTesti}{TRUE by default}
}
\value{
l'output etc
}
\description{
Assegna una sentiment per ogni testo in \code{text}
}
\details{
aggiungere dettagli qui
}
\note{
%% ~~further notes~~
}
\examples{

sentiment(c("ciao bella", "ciao", "good","casa", "farabutto!"))

}
\author{
Maddalena Branca, Livio Finos
}
\keyword{~kwd1}
\keyword{~kwd2}
