% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryTolower.R
\name{tryTolower}
\alias{tryTolower}
\title{tryTolower}
\usage{
tryTolower(testo, ifErrorReturnText = FALSE)
}
\arguments{
\item{testo}{a vector of characters to be normalized}
}
\value{
a vector of normalized characters
}
\description{
finds capital letters and replaces them with the corresponding lower ones
}
\details{
\code{tryTolower} finds capital letters in \code{testo} and replaces them with the corresponding lower ones
}
\examples{
testo<-c("Anna", "B")
tryTolower(testo)

}
\author{
Livio Finos
}
