% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizzaTesti.R
\name{normalizzaTesti}
\alias{normalizzaTesti}
\alias{normalizzacaratteri}
\alias{normalizzaemote}
\alias{normalizzahtml}
\alias{normalizzaslang}
\alias{normalizzapunteggiatura}
\alias{tryTolower}
\alias{itastopwords}
\alias{removeStopwords}
\alias{preprocessingEncoding}
\title{Varie funzioni di normalizzazione del testo}
\usage{
normalizzaTesti(testo, tolower = TRUE, normalizzahtml = TRUE,
  normalizzacaratteri = TRUE, normalizzaemote = TRUE,
  normalizzaEmoticons = TRUE, normalizzapunteggiatura = TRUE,
  normalizzaslang = TRUE, fixed = TRUE, perl = TRUE,
  preprocessingEncoding = TRUE, encoding = "UTF-8", sub = "",
  contaStringhe = c("\\\\?", "\\\\!", "@", "#", "(\\200|euro)",
  "(\\\\$|dollar)", "SUPPRESSEDTEXT"), suppressInvalidTexts = TRUE,
  verbatim = TRUE, remove = TRUE)
}
\arguments{
\item{testo}{character vector of texts}

\item{tolower}{\code{TRUE} by default}

\item{normalizzahtml}{\code{TRUE} by default}

\item{normalizzacaratteri}{\code{TRUE} by default}

\item{normalizzaemote}{\code{TRUE} by default}

\item{normalizzaEmoticons}{\code{TRUE} by default}

\item{normalizzapunteggiatura}{\code{TRUE} by default}

\item{normalizzaslang}{\code{TRUE} by default}

\item{fixed}{vedi \code{\link[base:gsub]{base:gsub}}. Preferibilmente non
usare l'opzione.}

\item{perl}{vedi \code{\link[base:gsub]{base:gsub}}. Preferibilmente non
usare l'opzione.}

\item{preprocessingEncoding}{logical}

\item{encoding}{\code{"UTF-8"} default. Se \code{FALSE} evita la
conversione.}

\item{sub}{character string. If not NA it is used to replace any
non-convertible bytes in the input. See also parameter \code{sub} in
function \code{iconv}.}

\item{contaStringhe}{stringhe da contare nei documenti. Default: \code{
c("\\?","\\!","#","@", "(€|euro)","(\\$|dollar)","SUPPRESSEDTEXT")}}

\item{suppressInvalidTexts}{Sostituisce con \code{"SUPPRESSEDTEXT"} le
stringhe con mutibyte non valida (che produrrebbero verosimilmente errori
nelle successive normalizzazioni). Default \code{TRUE}.}

\item{verbatim}{Mostra statitiche durante il processo. Default \code{TRUE}}

\item{remove}{\code{TRUE} by default. Possibily, a vector of stopwords to be
removed.}

\item{removeUnderscore}{rimuovere gli underscore?}

\item{ifErrorReturnText}{what to return for tests with a wrong encoding.}

\item{stopwords}{Lista di parole da escludere dall'analisi. A list of words
to be excluded from the process. \code{itastopwords} by default.}
}
\value{
Per \code{normalizzaTesti} l'output e' il vettore di testi
normalizzati.  La tabella dei conteggi specificati in \code{contaStringhe}
e' assegnato come tabella \code{counts} tra gli \code{attributes} del
vettore stesso.

Per tutte le altre funzioni, l'output e' un \code{vector} della stessa
lunghezza di \code{testo} ma con testi normalizzati.
}
\description{
Varie funzioni di normalizzazione del testo
}
\details{
\code{itastopwords} e' una lista di stopwords italiane.
}
\note{
%% ~~further notes~~
}
\examples{

testoNorm <- normalizzaTesti(c('ciao bella!','www.associazionerospo.org','noooo, che grandeeeeee!!!!!','mitticooo', 'mai possibile?!?!'))
testoNorm
attr(testoNorm,"counts")

}
\author{
Dario Solari, Livio Finos, Maddalena Branca
}
\keyword{~kwd1}
\keyword{~kwd2}
