% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{sentiment}
\alias{sentiment}
\alias{sentimentVocabularies}
\alias{vocabolariMadda}
\alias{vocabolarioMattivio}
\title{Performs sentiment analysis}
\usage{
sentiment(
  text,
  algorithm = "Mattivio",
  vocabularies = NULL,
  normalizzaTesti = TRUE,
  get_labels = TRUE,
  ...
)
}
\arguments{
\item{text}{vector of texts}

\item{algorithm}{\code{"Mattivio"} (default), \code{"Maddalena"} or a function returning the score.}

\item{vocabularies}{\code{vocabolarioMattivio} (default if \code{algorithm == "Mattivio"}), \code{vocabolariMadda} (default if \code{algorithm == "Maddalena"}) or an object used by the algorithm}

\item{normalizzaTesti}{TRUE by default}

\item{get_labels}{(TRUE) if FALSE the quantitavive score is given, if TRUE (the default) the labels -1,0,+1 are provied (i.e. \code{sign(score)})}
}
\value{
l'output etc
}
\description{
Assegna una sentiment per ogni testo in \code{text}
}
\details{
aggiungere dettagli qui
}
\note{
%% ~~further notes~~
}
\examples{

sentiment(c("ciao bella", "mi piaci", "wow!!","good","casa", "farabutto!","ti odio"))

}
\author{
Maddalena Branca, Mattia Da Pont, Livio Finos
}
\keyword{~kwd1}
\keyword{~kwd2}
