% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlExtract.R
\name{urlExtract}
\alias{urlExtract}
\alias{shorturl2url}
\alias{patternExtract}
\title{Estrazione di regular expression (e quindi users, hashtag) e shorturl}
\usage{
urlExtract(testo, id = names(testo))
}
\arguments{
\item{testo}{Vettore (eventualmente con nomi) di testi contenenti shorturl.}

\item{id}{se \code{testo} e' un vettore con nomi, questi vengono presi come
id. In caso contrario, gli id sono numeri progressivi da 1 a length(testo)}

\item{pattern}{stringa di testo da cercare ed estrarre. \code{"@\\w+"}
(default) estrae i riferimenti ad uno user nei tweets. \code{"#\\w+"} estrae
gli hashtag.}
}
\value{
\code{patternExtract} restituisce un \code{data.frame} con
colonne:id, pattern

\code{urlExtract} restituisce un \code{data.frame} con colonne:id, shorturl
e url
}
\description{
\code{patternExtract} estrae i pattern contenuti in in \code{testo}.
\code{urlExtract} estrae e converte gli shorturl contenuti in \code{testo}
in url. \code{shorturl2url} sostituisce gli shorturl contenuti in
\code{testo} in url.
}
\note{
%% ~~further notes~~
}
\examples{

 \dontrun{
 testo=c("Influenza Vaccination | ONS - Oncology Nursing Society http://t.co/924sRKGBU9 See All http://t.co/dbtPJRMl00,See All http://t.co/dbtPJRMl00")
 shorturl2url(testo,id=names(testo))
urls=urlExtract(testo)
patternExtract(c("@luca @paolo: buon giorno!", "@matteo: a te!"), pattern="@\\w+")
}

}
\author{
Dario Solari, Livio Finos
}
\keyword{~kwd1}
\keyword{~kwd2}
